finished SetUp but doesnt work correctly

+++ groceryShoppingList/app/src/main/java/com/example/groceryshoppinglist/Item.java
@@ -11,7 +11,7 @@ public class Item {
     @PrimaryKey(autoGenerate = true)
     @NonNull
     @ColumnInfo(name = "item_ID")
-    private Integer mItemID;
+    private Long mItemID;
 
     @NonNull
     @ColumnInfo(name = "item_Name")
@@ -30,7 +30,7 @@ public class Item {
     private String mCategory;
 
     // Constructor
-    public Item(@NonNull Integer itemID, @NonNull String item, @NonNull String quantity, @NonNull Integer listID, @NonNull String category) {
+    public Item(@NonNull Long itemID, @NonNull String item, @NonNull String quantity, @NonNull Integer listID, @NonNull String category) {
         this.mItemID = itemID;
         this.mItem = item;
         this.mQuantity = quantity;
@@ -53,6 +53,6 @@ public class Item {
     public Integer getListID(){return this.mListID;}
     public void setListID(Integer listID){this.mListID = listID;}
 
-    public Integer getItemID(){return this.mItemID;}
+    public Long getItemID(){return this.mItemID;}
 
 }
